+++ src/backend/app/projects/project_routes.py
@@ -518,7 +518,7 @@ async def task_split(
     features = await run_in_threadpool(
         lambda: split_by_sql(
             merged_boundary,
-            settings.FMTM_DB_URL.unicode_string(),
+            settings.FMTM_DB_URL,
             num_buildings=no_of_buildings,
             osm_extract=parsed_extract,
         )
@@ -568,7 +568,7 @@ async def preview_split_by_square(
 
     return split_by_square(
         boundary_featcol,
-        settings.FMTM_DB_URL.unicode_string(),
+        settings.FMTM_DB_URL,
         meters=dimension_meters,
         osm_extract=parsed_extract,
     )
@@ -806,15 +806,11 @@ async def generate_files(
     project = project_user_dict.get("project")
     project_id = project.id
     new_geom_type = project.new_geom_type
-
-    # Project requirement if they need to use odk-collect
     use_odk_collect = project.use_odk_collect or False
+    form_name = f"FMTM_Project_{project.id}"
+    project_contains_existing_feature = True if combined_features_count else False
 
     log.debug(f"Generating additional files for project: {project.id}")
-
-    form_name = f"FMTM_Project_{project.id}"
-
-    project_contains_existing_feature = True if combined_features_count else False
 
     # Validate uploaded form
     await central_crud.validate_and_update_user_xlsform(
@@ -1005,7 +1001,7 @@ async def upload_project_task_boundaries(
 @router.post("", response_model=project_schemas.ProjectOut)
 async def create_project(
     project_info: project_schemas.ProjectIn,
-    org_user_dict: Annotated[AuthUser, Depends(org_admin)],
+    org_user_dict: Annotated[OrgUserDict, Depends(org_admin)],
     db: Annotated[Connection, Depends(db_conn)],
 ):
     """Create a project in ODK Central and the local database.