+++ src/frontend/src/components/createnewproject/SelectForm.tsx
@@ -83,7 +83,9 @@ const SelectForm = ({ flag, _geojsonFile, xlsFormFile, setXlsFormFile }) => {
 
   useEffect(() => {
     if (xlsFormFile && !customFileValidity) {
-      dispatch(ValidateCustomForm(`${VITE_API_URL}/central/validate-form`, xlsFormFile?.file));
+      dispatch(
+        ValidateCustomForm(`${VITE_API_URL}/central/validate-form`, xlsFormFile?.file, projectDetails.use_odk_collect),
+      );
     }
   }, [xlsFormFile]);